Fix some pedantic errors

+++ diagnostics.h
@@ -16,9 +16,8 @@ enum diagnostics_type_e {
 
 void diagnostics_initialize(void);
 void diagnostics_type(enum diagnostics_type_e);
-void diagnostics_write_string(const char *s, size_t len);
-void diagnostics_write_string_escape_non_printable(const char *data,
-                                                   size_t len);
+void diagnostics_write_string(const char *s, int len);
+void diagnostics_write_string_escape_non_printable(const char *data, int len);
 void diagnostics_write_int(int n);
 void diagnostics_flush(void);
